requirements.txt
feature_transform==0.3.0
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
pillow
pytorch-lightning
torcharc==1.2.0
-e .
